+++ src/index.ts
@@ -1,5 +1,6 @@
 import { createQueryBuilder } from './query-builder';
 import { createAggregationBuilder } from './aggregation-builder';
 
-export const query = <T>() => createQueryBuilder<T>();
+export const query = <T>(includeQuery: boolean = true) =>
+  createQueryBuilder<T>({ _includeQuery: includeQuery });
 export const aggregations = <T>() => createAggregationBuilder<T>();